app/requirements.txt
alembic==1.11.1
anyio==3.7.1
backports.zoneinfo==0.2.1
certifi==2023.5.7
click==8.1.3
colorama==0.4.6
exceptiongroup==1.1.2
fastapi==0.99.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
Mako==1.2.4
MarkupSafe==2.1.3
psycopg==3.1.9
psycopg2-binary==2.9.6
pydantic==1.10.11
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.18
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.22.0
zipp==3.16.0
